BetterBehaviourEditor shouldn't leak anymore.
There was a bug that happens when you switch back to inherit from MonoBehaviour after previously inheriting BetterBehaviour.
If you do that and have your gameObject selected, you'd start to get null refs as if something's trying to access stuff from the old BetterBehaviourEditor which should have been destroyed by now.
But it seems not. So I added this check to make sure the TypedTarget (target as BetterBehaviour) is not null. Checking for target alone, it wasn't null.