Fixed wrong HelpBox size in GUIWrapper
This has been an ongoing-issue, when I first began implementing this, I couldn't get the 'right' height of the help box content. I ended up with ridiculously large numbers so I divided on 11 to get a decent height.
But now it seems the height is small when drawing a HelpBox, removing the division fixed the issue. Idk if it's something I fixed, or Unity in their 4.5