NEW: Added a RequireFromThis and RequireFromChildren attributes
This pretty much makes the assertions in Awake/Start/etc redundant. It's also more flexible than RequireComponent.
RequireFromThis requires your reference (Component) from the gameObject your script is attached to. You could tell it to add the component if not found.
RequireFromChildren requires it from the children gos instead. You could give it a path to add the component to if it wasn't found.