Made TransformEditor less error-prone
I should mention that the errors that sometimes get caused seemingly from TransformEditor is *not* because of TransformEditor. The errors are internal, complains about the creation of ScriptableObjects. The console gets bloated (999+ errors) I had this very same issue with NGUI's TransformEditor when working previously with NGUI. The error occurs randomly. But it's harmless. Restart Unity and it's no longer there.